Updated LICENSE and README.md
The license hadn't been updated since 2021 and I thought I'd update the year to a range of years; I plan for Custodia to have many releases as we move into the future, and I believe a range of years in the license best fits this situation. I also added my middle initial to my name within the license.

As far as the README.md is concerned, I have a writing addiction and constantly think of ways to improve and elaborate further upon prior statements.

This needlessly lengthy commit note is a perfect example of this trait of mine :frog: